added the automatic detection mechanism
This is the Python script for detecting the kite automatically, when a kite is present. It is based on detecting the difference in the RGB code, more specifically in the red channel (R) between the original image and a smoothed image using Gaussian smoothing. It may be better to split the function module from the execution module, which I will take care of in the following days, as well as commenting the modules to improve readability.